refactored SparkSegmentTarPushJobRunner as per the common code abstraction.
