+++ pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-spark-2.4/src/main/java/org/apache/pinot/plugin/ingestion/batch/spark/SparkSegmentUriPushJobRunner.java
@@ -37,8 +37,12 @@ import org.apache.spark.api.java.function.VoidFunction;
 
 public class SparkSegmentUriPushJobRunner extends AbstractSparkSegmentUriPushJobRunner {
 
+  public SparkSegmentUriPushJobRunner() {
+    super();
+  }
+
   public SparkSegmentUriPushJobRunner(SegmentGenerationJobSpec spec) {
-    super.init(spec);
+    super(spec);
   }
 
   @Override
